requirements.txt
requests
python-dotenv
PyPDF2
colpali-engine
requests
python-dotenv
PyPDF2
colpali-engine
PyMuPDF
Pillow
transformers==4.51.0
torch==2.4.1
torchvision==0.19.1
lancedb
rank_bm25
fuzzywuzzy
python-Levenshtein
torchaudio
sentencepiece
accelerate
docling
cachetools
numpy
networkx
matplotlib
psutil
httpx
scikit-learn
pandas
sentence_transformers
rerankers
nltk
vllm>=0.3.0; sys_platform != "darwin"  # vLLM doesn't support macOS
